third_party/WebKit/LayoutTests/inspector/sources/debugger/internal-script-flag-for-injected-script.html

<html>
<head>
<script src="../../../http/tests/inspector/inspector-test.js"></script>
<script src="../../../http/tests/inspector/console-test.js"></script>

<script>
function newWorld()
{
    testRunner.evaluateScriptInIsolatedWorld(239, "console.log(42);\n//# sourceURL=239.js");
}

function test()
{
    InspectorTest.debuggerModel.addEventListener(WebInspector.DebuggerModel.Events.ParsedScriptSource, step2);
    InspectorTest.evaluateInPage("newWorld()");

    var expectedScriptParsed = 2;
    function step2(event)
    {
        var script = event.data;
        InspectorTest.addResult(script._isInternalScript ? "Internal script" : "Not internal script");
        --expectedScriptParsed;
        if (!expectedScriptParsed)
            InspectorTest.completeTest();
    }
}
</script>
</head>
<body onload="runTest()">
<p>
Tests that injected script has isInternalScript flag.
</p>
</body>
</html>
